+++ SQLiteStudio3/guiSQLiteStudio/datagrid/sqlquerymodel.cpp
@@ -180,9 +180,12 @@ void SqlQueryModel::internalExecutionStopped()
     emit loadingEnded(false);
 }
 
-void SqlQueryModel::interrupt()
+void SqlQueryModel::interrupt(bool sync)
 {
-    queryExecutor->interrupt();
+    if (sync)
+        queryExecutor->interruptSync();
+    else
+        queryExecutor->interrupt();
 }
 
 qint64 SqlQueryModel::getExecutionTime()
@@ -1414,6 +1417,12 @@ void SqlQueryModel::handleExecFinished(SqlQueryPtr results)
 
     requiredDbAttaches = queryExecutor->getRequiredDbAttaches();
     reloadAvailable = true;
+
+    if (queryExecutor->isInterrupted())
+    {
+        emit executionFailed(tr("Error while executing SQL query on database '%1': %2").arg(db->getName(), "execution interrupted"));
+        return;
+    }
 
     emit loadingEnded(true);
     restoreNumbersToQueryExecutor();