+++ SQLiteStudio3/guiSQLiteStudio/mdichild.cpp
@@ -75,3 +75,8 @@ Db* MdiChild::getAssociatedDb() const
 void MdiChild::dbClosedFinalCleanup()
 {
 }
+
+bool MdiChild::isWindowClosingBlocked() const
+{
+    return false;
+}